tweaked health bars to use 64-bit numbers (ExpPD)

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ui/BossHealthBar.java
@@ -1,12 +1,16 @@
 /*
+ *
  * Pixel Dungeon
  * Copyright (C) 2012-2015 Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2023 Evan Debenham
+ * Copyright (C) 2014-2024 Evan Debenham
  *
  * Experienced Pixel Dungeon
- * Copyright (C) 2019-2020 Trashbox Bobylev
+ * Copyright (C) 2019-2024 Trashbox Bobylev
+ *
+ * Extended Experienced Pixel Dungeon
+ * Copyright (C) 2023-2024 John Nollas
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,6 +24,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
  */
 
 package com.shatteredpixel.shatteredpixeldungeon.ui;
@@ -176,9 +181,9 @@ public class BossHealthBar extends Component {
 				long shield = boss.shielding();
 				long max = boss.HT;
 
-				hp.scale.x = Math.max( 0, (health-shield)/(float)max);
-				shieldedHP.scale.x = health/(float)max;
-				rawShielding.scale.x = shield/(float)max;
+				hp.scale.x = (float) Math.max( 0, (health-shield)/(double)max);
+				shieldedHP.scale.x = (float) (health/(double)max);
+				rawShielding.scale.x = (float) (shield/(double)max);
 
 				if (hp.scale.x < 0.25f) bleed( true );
